use chatMessaging template in chatroomPage
