--- webapp/src/pages/ChatroomPage/ChatroomPage.css
@@ -1,39 +1,0 @@
-.chatroom {
-  padding: var(--padding-top-bottom) var(--padding-sides);
-  height: 100vh;
-  background-color: var(--color-background);
-  color: var(--color-light);
-  display: grid;
-  gap: 1rem;
-  grid-template-areas:
-    'header'
-    'messages'
-    'message-input';
-  grid-template-rows: auto 1fr auto;
-  grid-template-columns: 1fr;
-}
-
-.chatroom-loading {
-  grid-column: 1/-1;
-  grid-row: 1/-1;
-  align-self: center;
-  justify-self: center;
-}
-
-.chatroom-header {
-  grid-area: header;
-}
-
-.chatroom-messages {
-  grid-area: messages;
-  overflow-y: scroll;
-  overflow-x: hidden;
-
-  /* Scroll to the bottom of the messages list using CSS - https://stackoverflow.com/a/72644230 */
-  display: flex;
-  flex-direction: column-reverse;
-}
-
-.chatroom-message-input {
-  grid-area: message-input;
-}
