--- webapp/src/pages/ChatroomPage/components/ChatroomMessageInput/ChatroomMessageInput.css
@@ -1,56 +1,0 @@
-.chatroom-message-input-form {
-  width: 100%;
-  display: grid;
-  gap: 0.5rem;
-  grid-template-columns: 1fr;
-  grid-template-rows: 1fr auto;
-  grid-template-areas: 'message-input';
-}
-
-.chatroom-message-input-grow-wrap {
-  grid-area: message-input;
-  /* easy way to plop the elements on top of each other and have them both sized based on the tallest one's height */
-  display: grid;
-  grid-template-columns: 1fr auto;
-  gap: 1rem;
-}
-
-.chatroom-message-input-grow-wrap::after {
-  /* Note the weird space! Needed to preventy jumpy behavior */
-  content: attr(data-replicated-value) ' ';
-
-  /* This is how textarea text behaves */
-  white-space: pre-wrap;
-
-  /* Hidden from view, clicks, and screen readers */
-  visibility: hidden;
-}
-
-.chatroom-message-input-text {
-  /* You could leave this, but after a user resizes, then it ruins the auto sizing */
-  resize: none;
-}
-
-.chatroom-message-input-text::placeholder {
-  color: var(--color-light);
-  opacity: 1;
-}
-
-.chatroom-message-input-grow-wrap::after,
-.chatroom-message-input-text {
-  /* Identical styling required!! */
-  outline: none;
-  border: none;
-  margin: 0;
-  min-height: 3rem;
-  max-height: 12rem;
-  overflow-y: scroll;
-  overflow-x: hidden;
-  background-color: var(--color-dark);
-  color: var(--color-light);
-  padding: 0.8rem 0.6rem;
-  border-radius: 0.4rem;
-
-  /* Place on top of each other */
-  grid-area: 1 / 1 / 2 / 2;
-}
