--- webapp/src/pages/ChatroomPage/components/index.ts
@@ -1,2 +1,0 @@
-export { default as ChatroomMessageInput } from './ChatroomMessageInput/ChatroomMessageInput';
-export { default as ChatroomMessages } from './ChatroomMessages/ChatroomMessages';
